For WeatherDataAPI, added new constructor and updateWeatherData()
The existing constructor takes a JSONObject as a parameter, therefore
requiring that the API already have been called. The new constructor
takes a String inputLocation as a parameter, and thus allows for all API
calls (both initial and subsequent) to flow through WeatherDataAPI
objects.
The updateWeatherData() method re-calls the API using the object's
inputLocation field, and updates all fields accordingly. If our first
constructor was used for object creation, inputLocation is identical to
the address field. If the new constructor was used, inputLocation is the
location name as entered by the user (and passed as the argument to the
constructor).