Added handling for "Bad API Request"
Invalid locations (e.q., misspelling of a city name) can result in a Bad
API Request to the Visual Crossing API. We now check for this while
creating a JSONObject to be returned in
WeatherService.fetchWeatherData(). We also add a "validAddress" field to
the JSONObject that is false if a Bad API Request occurred. The
WeatherDataAPI class also now has a corresponding Boolean field for that
validAddress. WeatherDataAPI also has a public isValid() method that
indicates if a Bad API Request occurred.