Additional WeatherService error handling; Updated getter names
WeatherService previously did not account for the following error from
Visual Crossing: "Bad API Request:A location must be specified". This
would produce an error call stack in the console when the user tried to
click the "Go" button when the text box was empty. The error is now
accounted for.

Updated some of WeatherDataAPI's getter names to be more descriptive.
getAddress() changed to getResolvedAddress(). getTime() changed to
getCurrentTime(). WeatherDataAPI.getHour() was removed because getTime()
removes a ZonedDateTime object that has its own getHour() method.

Commented out some out.println lines in FXMLController that were used
for testing.