Miscellaneous FXMLController Changes (see body of commit message)
Changes include:
 - In setSunTimes(), changed DateTimeFormatter pattern from "HH:mm" to
"h:mm a". The single "h" makes it use one digit when possible (4:12
instead of 04:12). The "a" adds AM/PM. That the "h" is lowercase makes
it use 12-hour clock (4:12PM instead of 16:12PM).
 - Removed unused "int startingHour" parameter from several
FXMLController methods.
 - Added "private" modifier to "date" and "currentHour" fields
 - Removed initializeHours()'s helper function timeToString() --> found
a way to do it more concisely with DateTimeFormatter within
initializeHours()
 - Changed repeated if statements to single if-else statements in
directionToString() and initializeUV()
 - Commented out all appearances of viewHour and the shiftHourHandler()
method (both unused)
 - Misc. stylistic consistency/changes (e.g., tabs, empty lines, braces)